+++ local_test_settings.py
@@ -9,3 +9,7 @@ INSTALLED_APPS += (
 
 CELERY_ALWAYS_EAGER = True
 BROKER_BACKEND = 'memory'
+
+PERSONALISATION_SEGMENTS_ADAPTER = (
+    'molo.surveys.adapters.SurveysSegmentsAdapter'
+)