+++ molo/surveys/models.py
@@ -44,11 +44,12 @@ from wagtailsurveys.models import AbstractFormField
 
 from .blocks import SkipLogicField, SkipState, SkipLogicStreamPanel
 from .forms import MoloSurveyForm, PersonalisableMoloSurveyForm
-from .rules import GroupMembershipRule, SurveySubmissionDataRule  # noqa
+from .rules import ArticleTagRule, GroupMembershipRule, SurveySubmissionDataRule  # noqa
 from .utils import SkipLogicPaginator
 
 
 SKIP = 'NA (Skipped)'
+
 
 # See docs: https://github.com/torchbox/wagtailsurveys
 SectionPage.subpage_types += ['surveys.MoloSurveyPage']
@@ -252,9 +253,15 @@ class MoloSurveyPage(
         When the last step is submitted correctly, the whole form is saved in
         the DB.
         """
-        paginator = SkipLogicPaginator(self.get_form_fields(), request.POST)
-
         session_key_data = 'survey_data-%s' % self.pk
+        survey_data = json.loads(request.session.get(session_key_data, '{}'))
+
+        paginator = SkipLogicPaginator(
+            self.get_form_fields(),
+            request.POST,
+            survey_data,
+        )
+
         is_last_step = False
         step_number = request.GET.get('p', 1)
 
@@ -275,12 +282,10 @@ class MoloSurveyPage(
 
             # Create a form only for submitted step
             prev_form_class = self.get_form_class_for_step(prev_step)
-            prev_form = prev_form_class(request.POST, page=self,
+            prev_form = prev_form_class(paginator.new_answers, page=self,
                                         user=request.user)
             if prev_form.is_valid():
                 # If data for step is valid, update the session
-                survey_data = json.loads(
-                    request.session.get(session_key_data, '{}'))
                 survey_data.update(prev_form.cleaned_data)
                 request.session[session_key_data] = json.dumps(
                     survey_data, cls=DjangoJSONEncoder)
@@ -392,6 +397,12 @@ class SkipLogicMixin(models.Model):
         )
 
     def choice_index(self, choice):
+        if self.field_type == 'checkbox':
+            # clean checkboxes have True/False
+            try:
+                return ['on', 'off'].index(choice)
+            except ValueError:
+                return [True, False].index(choice)
         return self.choices.split(',').index(choice)
 
     def next_action(self, choice):
@@ -416,7 +427,8 @@ class SkipLogicMixin(models.Model):
 
     def save(self, *args, **kwargs):
         self.choices = ','.join(
-            choice.value['choice'] for choice in self.skip_logic
+            choice.value['choice'].replace(',', u'\u201A')
+            for choice in self.skip_logic
         )
         return super(SkipLogicMixin, self).save(*args, **kwargs)